feat: display larc sections
